+++ Api-Gateway/index.js
@@ -5,12 +5,15 @@ require("dotenv").config();
 const https = require("https");
 const path = require("path");
 const fs = require("fs");
-
+const cors = require("cors");
+const swaggerjsdoc = require("swagger-jsdoc");
+const swaggerui = require("swagger-ui-express");
 const routes = {
   "/orders": process.env.ORDERS,
   "/profile": process.env.PROFILE,
   "/users": process.env.USERS,
 };
+app.use(cors());
 app.get("/", (req, res) => {
   res.send("Home");
 });
@@ -39,6 +42,29 @@ for (const route in routes) {
 //   console.log("Orders. Port :- 3000");
 // });
 
+const options = {
+  definition: {
+    openapi: "3.0.0",
+    info: {
+      title: "Microservices ",
+      version: "1.0.0",
+    },
+
+    servers: [{ url: "http://localhost:8000" }],
+  },
+  apis: [
+    "./swaggers/authentication/*.js",
+    "./swaggers/profile/*.js",
+    "./swaggers/tasks/*.js",
+    "./swaggers/admin/*.js",
+  ],
+};
+const spacs = swaggerjsdoc(options);
+app.use("/api-docs", swaggerui.serve, swaggerui.setup(spacs));
+app.get("/docs.json", (req, res) => {
+  res.setHeader("Content-Type", "application/json");
+  res.send(spacs);
+});
 app.listen(process.env.PORT, () => {
   console.log(` Port :- ${process.env.PORT}`);
 });
